% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{calculateSummaryStatistics}
\alias{calculateSummaryStatistics}
\title{Summary Statistics}
\usage{
calculateSummaryStatistics(simulation, strict = TRUE)
}
\arguments{
\item{simulation}{Simulation output of the class "Phylosim"}

\item{strict}{Logical. If true the output of the function will be NA if at least one summary statistic could not be calculated.}
}
\value{
A list containing the summary statistics.
}
\description{
Calculates the summary satistics for a simulation of class PhyloSim.
}
\details{
TODO: implement
}
\examples{

simu <- runSimulation(createCompletePar(x=50, y=50, runs=1000, specRate=2))
summStat <- calculateSummaryStatistics(simu)
}
