% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{callModel}
\alias{callModel}
\title{Core phylosim model}
\usage{
callModel(x, y, dispersal, runs, specRate, negativeDens, positiveDens, env,
  neutral, mort, mortStrength, repro, dispersalCutoff, nDensCutoff,
  pDensCutoff, seed, envStrength, nDDStrength, pDDStrength, fission, redQueen,
  redQueenStrength, protracted, airmatR, soilmatR, prunePhylogeny,
  envNicheWidth, nDDNicheWidth, pDDNicheWidth)
}
\description{
Core phylosim model
}
