%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{checkConsistency.CodonSequence}
\alias{checkConsistency.CodonSequence}
\alias{CodonSequence.checkConsistency}
\alias{checkConsistency,CodonSequence-method}

 \title{Check object consistency}

 \description{
		Check object consistency.
 }

 \usage{
\method{checkConsistency}{CodonSequence}(this, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{...}{Not used.}
 }


 \value{
		Returns an invisible TRUE if no inconsistencies found in the object, throws
		an error otherwise.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{CodonSequence}}.
 }
