%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{checkConsistency.QMatrix}
\alias{checkConsistency.QMatrix}
\alias{QMatrix.checkConsistency}
\alias{checkConsistency,QMatrix-method}

 \title{Check object consistency}

 \description{
		Check object consistency.
 }

 \usage{
\method{checkConsistency}{QMatrix}(this, check.process=TRUE, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{check.process}{Check the associated process object.}
       \item{...}{Not used.}
 }


 \value{
		Returns an invisible TRUE if no inconsistencies found in the object, throws
		an error otherwise.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{QMatrix}}.
 }
