% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{cleanUpChunks}
\alias{cleanUpChunks}
\title{Clean up parameter and summary chunks}
\usage{
cleanUpChunks(dir = ".")
}
\arguments{
\item{dir}{Directory to which the results of chunkwiseCalculateSumaries and runPhylosimAnalysis were saved.}
}
\description{
Reads in summary files created with chunkwiseCalculateSummaries, and parameter files created with runPhylosimAnalysis and removes crashed simulations (those where summaries == NA). The resulting cleaned-up chunks are saved to the subfolder "cleaned_up".
}
\author{
Tankred Ott
}
