% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{combineCunks}
\alias{combineCunks}
\title{combine chunks}
\usage{
combineCunks(dir = ".", prefix)
}
\arguments{
\item{dir}{Directory. Location of the files}

\item{prefix}{Common file prefix. Possible to use RegEx like syntax here.}
}
\description{
Combines files with a common prefix to a combined files and saves the files as combined_PREFIX.
}
\author{
Tankred Ott
}
