% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Sequence Concatenation}
\usage{
concatenate(x, discard = NULL)
}
\arguments{
\item{x}{A \code{\link{raw}} vector storing the molecular traits (single
nucleotides, gaps, and so on).}

\item{discard}{A vector of characters to exclude from the results (default:
\code{NULL}, meaning nothing is excluded).}
}
\value{
A vector of character strings.
}
\description{
A function to transform a \code{\link{raw}} matrix, used to
store sets of single nucleotides, into a vector of strings, optionally
removing specific value(s).
}
\details{
That function concatenates the rows of a \code{\link{raw}} matrix,
which is used for storing the molecular trait values, into a vector of
strings. If argument \code{discard} is given a set of values, they will be
excluded from the resulting strings.
}
\examples{
## Define a raw vector for storing nuceotide values:
c(Sequence_1 = "ATCG-TTTCG--CCCCA--TTA--TTAA-GTAA-GTAATCTTTCA",
  Sequence_2 = "TTGGCTTCC--TC--CAT-TTC--TTCA-GT-ACG-ATTCTTTTA",
  Sequence_3 = "TTCG-TACC-T-T---A-ATAA--T-AA-GTCTTGTAATCGTTCA") \%>\%
  sapply(charToRaw) \%>\%
  t -> sqn

## Display the raw sequence:
sqn

## Transforming the sequence to character strings
concatenate(sqn)

## Transforming the sequence to character strings without the gaps:
concatenate(sqn, discard="-")

## Clean-up:
rm(sqn)

}
\author{
\packageAuthor{PhyloSim}
}
