%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{detachSeqFromNode.PhyloSim}
\alias{detachSeqFromNode.PhyloSim}
\alias{PhyloSim.detachSeqFromNode}
\alias{detachSeqFromNode,PhyloSim-method}

 \title{Detach a Sequence object from a given node of a phylo object aggregated by a PhyloSim object}

 \description{
	Detach a Sequence object from a given node of a phylo object aggregated by a PhyloSim object.

	This method is mainly used internally.

 }

 \usage{
\method{detachSeqFromNode}{PhyloSim}(this, node=NA, ...)
}

 \arguments{
 	\item{this}{A PhyloSim object.}
	\item{node}{Node identifier.}
 	\item{...}{Not used.}
 }

 \value{
 	The PhyloSim object (invisible).
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{PhyloSim}}.
 }
