% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_idx.R
\name{dst_idx}
\alias{dst_idx}
\title{Distance Vector Indexing}
\usage{
dst_idx(n, i, j)
}
\arguments{
\item{n}{An integer giving the number of objects from which the distances
were calculated.}

\item{i}{An integer giving the index of the first object.}

\item{j}{An integer giving the index of the second object (can be omitted,
see details).}
}
\value{
The index of the distances in the vector of distance.
}
\description{
Function to calculate the indices of a distance vector that contains the
pairwise distance between two objects.
}
\details{
A distance vector contains the lower triangle of a pairwise distance
matrix concatenated in a column by column manner. While storing distance in
that way saves substantial space (i.e., by avoiding the repeated upper
triangle values and the all-zero diagonal values), accessing the values is
somewhat more tedious than for a square matrix, for which the R language
provide extraction and replacement operators.

When argument \code{j} is omitted, all the indices between object \code{i}
and the other observations are returned. In that case, argument \code{i}
accepts only a single value.
}
\examples{
### Here...




}
\author{
\packageAuthor{PhyloSim}
}
