%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{enableVirtual.PSRoot}
\alias{enableVirtual.PSRoot}
\alias{PSRoot.enableVirtual}
\alias{enableVirtual,PSRoot-method}

 \title{Enable the use of virtual fields for a given object}

 \description{
	Enable the use of virtual fields for a given object.
	R.oo disables the virtual field feature inside get/set methods. This method can be used to re-enable virtual fields.
 }

 \usage{
\method{enableVirtual}{PSRoot}(this, ...)
}

 \arguments{
 	\item{this}{A PSRoot object.}
 	\item{...}{Not used.}
 }

 \value{
 	The PSRoot object.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{PSRoot}}.
 }
