% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedSpeciesRichness.R
\name{es}
\alias{es}
\title{Expected species richness}
\usage{
es(specRate, dimensions)
}
\arguments{
\item{specRate}{The speciation rate in speciations per generation}

\item{dimensions}{The length of the grid edge}
}
\value{
A float value for the expected species richness
}
\description{
Calculate the expected species richness in the equilibrium state as calculated by Hubbel (2001).
}
\details{
This function is only valid for neutral models!
}
\examples{
es(2, 100)
}
\references{
Hubbell, Stephen P. The unified neutral theory of biodiversity and biogeography (MPB-32). Vol. 32. Princeton University Press, 2001.
}

