% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{extractParams}
\alias{extractParams}
\title{Extract params}
\usage{
extractParams(parList, params, outFormat = "df")
}
\arguments{
\item{parList}{List of phylosim parameter objects}

\item{params}{Vector of parameter names to extract}

\item{outFormat}{Type of output that should be returned. "df" for dataframe or "list" for list.}
}
\description{
Extracts parameters from a list of PhyloSim parameter objects.
}
\author{
Tankred Ott
}
