% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircularOffsets.R
\name{getCircularOffsets}
\alias{getCircularOffsets}
\title{Compute toroidal neighbor shifts}
\usage{
getCircularOffsets(radius)
}
\arguments{
\item{radius}{Integer radius defining the circular neighborhood}
}
\value{
A data frame with columns \code{dx} and \code{dy} indicating x and y offsets
}
\description{
Computes relative coordinate offsets that define a circular neighborhood
of a given radius around a central cell, excluding the center (0, 0). Used in conspecific
neighborhood analysis on torus-expanded matrices.
}
\details{
Get circular neighborhood coordinate offsets
}
\examples{
getCircularOffsets(1)
getCircularOffsets(3)

}
