% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConNeigh.R
\name{getConNeigh}
\alias{getConNeigh}
\alias{getConNeigh.PhyloSim}
\alias{getConNeigh.PhylosimList}
\title{Get number of conspecific neighbors}
\usage{
getConNeigh(simu)

\method{getConNeigh}{PhyloSim}(simu)

\method{getConNeigh}{PhylosimList}(simu)
}
\arguments{
\item{simu}{Object of class \code{PhyloSim} or \code{PhylosimList}}
}
\value{
Input object with \code{conNeighMat} matrices added to all generations
}
\description{
Calculates the number of conspecific neighbors in a circular radius
for all generations based on torus-expanded \code{specMat}.
Automatically calls \code{\link{getTorus}} with \code{overwrite = TRUE},
which also prepares \code{idMat} and \code{mortMat} if missing.

The torus is automatically undone after computing neighborhood matrices.
}
\details{
Calculate conspecific neighborhood matrices
}
