%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getEventsAtSite.FastFieldDeletor}
\alias{getEventsAtSite.FastFieldDeletor}
\alias{FastFieldDeletor.getEventsAtSite}
\alias{getEventsAtSite,FastFieldDeletor-method}

 \title{Generate a deletion event object given the state of the target site}

 \description{
	Generate a deletion event object given the state of the target site.
 }

 \usage{
\method{getEventsAtSite}{FastFieldDeletor}(this, target.site, ...)
}

 \arguments{
       \item{this}{A FastFieldDeletor object.}
       \item{target.site}{A Site object.}
       \item{...}{Not used.}
 }

 \value{
 	A list of Event objects.
 }


 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	GeneralDeletor getEventsAtSite.GeneralDeletor
 }
