% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\alias{getID.PhyloSim}
\alias{getID.PhylosimList}
\title{Generate ID matrix for individuals}
\usage{
getID(simu)

\method{getID}{PhyloSim}(simu)

\method{getID}{PhylosimList}(simu)
}
\arguments{
\item{simu}{Object of class \code{PhyloSim} or \code{PhylosimList}}
}
\value{
Input object with \code{idMat} matrices added to each generation.  Also, adds the years of generations (e.g., \code{simu$Output$test$Output$`98`}.
}
\description{
Each individual gets a persistent ID. IDs are unique per run.
If the trait stays the same between generations, the ID persists. Otherwise, a new one is assigned.
}
\details{
Assign individual IDs across generations
}
