% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMortality.R
\name{getMortality}
\alias{getMortality}
\alias{getMortality.PhyloSim}
\alias{getMortality.PhylosimList}
\title{Get mortality of individuals}
\usage{
getMortality(simu)

\method{getMortality}{PhyloSim}(simu)

\method{getMortality}{PhylosimList}(simu)
}
\arguments{
\item{simu}{Object of class \code{PhyloSim} or \code{PhylosimList}}
}
\value{
Same object, with \code{mortMat} matrices added to \code{Output[[i]]}. Also, adds the years of generations (e.g., \code{simu$Output$test$Output$`98`}.
}
\description{
Compares each generation's \code{traitMat} to the next. If the trait
differs, the individual is considered dead. First generation gets NA. Generation
names are set using \code{Model$runs}.
}
\details{
Compute individual mortality across generations
}
