%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getRateParam.F81}
\alias{getRateParam.F81}
\alias{F81.getRateParam}
\alias{getRateParam,F81-method}

 \title{Forbidden action: the value of a rate parameters}

 \description{
       Forbidden action: the value of a rate parameters.

	This model has no rate parameters.
 }

 \usage{
\method{getRateParam}{F81}(this, name, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{name}{The name of the rate parameter.}
       \item{...}{Not used.}
 }

 \value{
	The object.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{F81}}.
 }
