% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpecies}
\alias{getSpecies}
\title{Get Recent Species}
\usage{
getSpecies(sim)
}
\arguments{
\item{sim}{Phylosim object}
}
\value{
List of vectors containing the species IDs.
}
\description{
Function to extract the recent species of a Phylosim object
}
\details{
The function extracts the IDs of the recent species of the grid in the Phylosim object.
}
\examples{
## Define a parameter set
par <- createCompletePar(x = 50, y = 50, dispersal = FALSE , runs = c(500,1000),
        density = 1, environment = 0.5, specRate = 1)

## Run the model
simu <- runSimulation(par)

## Extract species IDs
speciesIDs <- getSpecies(simu)

}
\author{
Tankred Ott
}
