% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpeciesIndices}
\alias{getSpeciesIndices}
\title{Get Species Indices}
\usage{
getSpeciesIndices(sim, speciesID, whichRun = 1)
}
\arguments{
\item{sim}{Phylosim object}

\item{speciesID}{Integer, species ID}

\item{whichRun}{Integer, specifying from which run the data should be collected.}
}
\value{
Vector of indices
}
\description{
Function to get the indices (positions) of a species on the Phylosim grid
}
\details{
The function extracts the environmental, competition and neutral trait of the species with speciesID from the Phylosim object. If there are multiple runs in
}
\author{
Tankred Ott
}
