% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpeciesTrait}
\alias{getSpeciesTrait}
\title{Get Species Trait}
\usage{
getSpeciesTrait(sim, speciesID, trait = "neutral", whichRun = 1)
}
\arguments{
\item{sim}{Phylosim object}

\item{speciesID}{Integer, species ID}

\item{trait}{String, determining which trait should be returned ("evironment", "competition", "neutral").}

\item{whichRun}{Integer, specifying from which run the data should be collected.}
}
\value{
Vector, containing the trait values
}
\description{
Function to extract the traits of a species from a Phylosim object.
}
\details{
The function extracts the environmental, competition or neutral trait of the species with speciesID from the Phylosim object. If there are multiple runs in whichRun determines which run should be used.
}
\examples{
## Define a parameter set
par <- createCompletePar(x = 50, y = 50, runs = c(500,1000),
        density = 1, environment = 0.5, specRate = 1)

## Run the model
simu <- runSimulation(par)

## Extract species IDs
speciesIDs <- getSpecies(simu)

## Get Species Traits of a species
envTraits <- PhyloSim:::getSpeciesTrait(simu, speciesIDs[[1]][1], "environment", 1)

hist(envTraits)

}
\author{
Tankred Ott
}
