% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpeciesTraits}
\alias{getSpeciesTraits}
\title{Get Species Traits}
\usage{
getSpeciesTraits(sim, speciesID, whichRun = 1)
}
\arguments{
\item{sim}{Phylosim object}

\item{speciesID}{Integer, species ID}

\item{whichRun}{Integer, specifying from which run the data should be collected.}
}
\value{
List of trait types containing vectors with trait values.
}
\description{
Function to extract the traits of a species from a Phylosim object.
}
\details{
The function extracts the environmental, competition and neutral trait of the species with speciesID from the Phylosim object. If there are multiple runs in whichRun determines which run should be used.
}
\examples{
## Define a parameter set
par <- createCompletePar(x = 50, y = 50, runs = c(500,1000),
        density = 1, environment = 0.5, specRate = 1)

## Run the model
simu <- runSimulation(par)

## Extract species IDs
speciesIDs <- getSpecies(simu)

## Get Species Traits of a species
traits <- getSpeciesTraits(simu, speciesIDs[[1]][1])

#hist(traits$environment)

}
\author{
Tankred Ott
}
