% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecTime.R
\name{getSpecTime}
\alias{getSpecTime}
\alias{getSpecTime.PhyloSim}
\alias{getSpecTime.PhylosimList}
\title{Get species richness time series}
\usage{
getSpecTime(runs, thinning_factor = NULL, ymax = NULL, plot = TRUE)

\method{getSpecTime}{PhyloSim}(runs, thinning_factor = NULL, ymax = NULL,
  plot = TRUE)

\method{getSpecTime}{PhylosimList}(runs, thinning_factor = NULL,
  ymax = NULL, plot = TRUE)
}
\arguments{
\item{runs}{A \code{PhyloSim} or \code{PhylosimList} object}

\item{thinning_factor}{Optional integer to reduce the number of time points for plotting (smootes the line). Turned off by default.}

\item{ymax}{Optional numeric; fixed y-axis max for plotting. Usefull for comparing different runs. Turned off by default.}

\item{plot}{Logical; whether to generate plots. Default is TRUE}
}
\value{
A \code{data.frame} (for single object) or list of data.frames (for list input)
}
\description{
Computes species richness per generation using \code{PhyloSim::specRich()}.
Can be visualized or just returned. Richness values can optionally be thinned.
}
\details{
Plot or extract species richness over time
}
