% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTorus.R
\name{getTorus}
\alias{getTorus}
\alias{getTorus.PhyloSim}
\alias{getTorus.PhylosimList}
\title{Toroidal expansion of simulation matrices}
\usage{
getTorus(simu, overwrite = FALSE, radius = NULL)

\method{getTorus}{PhyloSim}(simu, overwrite = TRUE, radius = NULL)

\method{getTorus}{PhylosimList}(simu, overwrite = FALSE, radius = NULL)
}
\arguments{
\item{simu}{Object of class \code{PhyloSim} or \code{PhylosimList}}

\item{overwrite}{Logical; if \code{TRUE}, original matrices will be overwritten}

\item{radius}{Integer; neighborhood radius. Defaults to \code{densityCut}}
}
\value{
Input object with torus-expanded matrices added or replaced
}
\description{
Adds torus-padding to all relevant matrices mainly to examine nieghbors also of edge cases. This can be done, because the simulation in c++ also uses toroidal boundaries. By default, generates \code{mortMat} and \code{idMat} if missing.
Names generations based on \code{Model$runs}.
}
\details{
Expand all matrices toroidally
}
\keyword{internal}
