% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTorus.R
\name{getTorusMatrix}
\alias{getTorusMatrix}
\title{Toroidal matrix expansion (helper)}
\usage{
getTorusMatrix(mat, r)
}
\arguments{
\item{mat}{Input matrix}

\item{r}{Radius to expand}
}
\value{
Expanded matrix with toroidal wrapping
}
\description{
Toroidal matrix expansion (helper)
}
\keyword{internal}
