% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPlots.R
\name{localPlots}
\alias{localPlots}
\title{Create Subplots}
\usage{
localPlots(simu, which.result = NULL, size, n, community = FALSE,
  plot = FALSE)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs to contain at least a species matrix ($specMat)}

\item{which.result}{Integer, determines which result should be used. This argument is only usefull if your 'runs' argument in \code{\link{createCompletePar}} contains more than one element. By default (NULL), the last result is used.}

\item{size}{A single value or a vector determining the size(s) of the generated subplots}

\item{n}{The number of subplots to be generated}

\item{community}{Logical, determining whether to generate a communiy table or not. default is FALSE}

\item{plot}{Logical, determining whether the results should be shown graphically. Default is FALSE.}
}
\value{
A list of subplots and (if coomunity = T) a community table with plots in rows and species in collumns
}
\description{
Generates subplots from a metacommunity, simulating field conditions
}

