% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molEvolMod.R
\name{molEvolMod}
\alias{molEvolMod}
\alias{read.mutationMat}
\title{Molecular Evolution Simulator}
\usage{
read.mutationMat(par)

molEvolMod(I, step, rho)
}
\arguments{
\item{par}{A list containing the parameters needed to construct the model's
transition intensity matrix.}

\item{I}{A 5 x 5 transition intensity matrix such as the one obtained from
function \code{read.mutationMat} (see details).}

\item{step}{The simulation time step (arbitrary units of time).}

\item{rho}{An evolution rate factor to be used on top of the transition
intensity matrix.}
}
\value{
\describe{
  \item{ \code{read.mutationMat} }{...}
  \item{ \code{molEvolMod} }{...}
}
}
\description{
Functions to simulate the evolution of DNA sequences following a
Markov process.
}
\details{
The workflow to simulate DNA sequence evolution consists in
generating a random sequence of nucleotides and gaps (A, T, C, G, -) and
simulating ...
}
\examples{
## Load the example of a configuration file provided with the
## package:
yaml.load_file(
  system.file(
    package = "PhyloSim",
    "extdata",
    "evolMod.yml"
  )
) -> dnaParam

## The configuration list is located in member `$DNA`
read.mutationMat(
  dnaParam$DNA 
) -> I

## The transition intensity matrix:
I

## Implement the molecular evolution model for a single nucleotide:
molEvolMod(I, 1, 1) -> em1

## Get the transition probability matrix as follows:
em1$getMt()

## A vector of raw as examples of initial traits:
tr <- charToRaw("-ACGT")

## Simulate molecular evolution from:
rawToChar(em1$evolve(tr[1L]))    ## a gap.
rawToChar(em1$evolve(tr[2L]))    ## an adenine base.
rawToChar(em1$evolve(tr[3L]))    ## a cytosine base.
rawToChar(em1$evolve(tr[4L]))    ## a guanine base.
rawToChar(em1$evolve(tr[5L]))    ## a thymine base.

## Recalculate the probabilities for a lower mean evolution rate (one tenth
## the previous one):
em1$recalculate(1, 0.1)

em1$getMt()        ## The recalculated transition probability matrix.

## Simulate molecular evolution from:
rawToChar(em1$evolve(tr[1L]))    ## a gap.
rawToChar(em1$evolve(tr[2L]))    ## an adenine base.
rawToChar(em1$evolve(tr[3L]))    ## a cytosine base.
rawToChar(em1$evolve(tr[4L]))    ## a guanine base.
rawToChar(em1$evolve(tr[5L]))    ## a thymine base.

## Base changes are now less probable.

## Clean up:
rm(em1, tr)

}
\author{
\packageAuthor{PhyloSim}
}
