%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{newAAMatrix.AminoAcidSubst}
\alias{newAAMatrix.AminoAcidSubst}
\alias{AminoAcidSubst.newAAMatrix}
\alias{newAAMatrix,AminoAcidSubst-method}

 \title{Undocumented method}

 \description{
	Undocumented method.
 }

 \usage{
\method{newAAMatrix}{AminoAcidSubst}(name=NA, paml.file=NA, equ.dist=NA, ...)
}

 \arguments{
 	\item{name}{Object name}
 	\item{paml.file}{PAML file.}
	\item{equ.dist}{Equilibrium distribution.}
 	\item{...}{Not used.}
 }

 \value{
 	A process object inheriting from AminoAcidSubst.
 }


 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{AminoAcidSubst}}.
 }
