%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{newMatrix.CodonSubst}
\alias{newMatrix.CodonSubst}
\alias{CodonSubst.newMatrix}
\alias{newMatrix,CodonSubst-method}

 \title{New codon substitution matrix from PAML file}

 \description{
	New codon substitution matrix from PAML file.
 }

 \usage{
\method{newMatrix}{CodonSubst}(name=NA, paml.file=NA, equ.dist=NA, ...)
}

 \arguments{
 	\item{name}{Object name}
 	\item{paml.file}{PAML file.}
	\item{equ.dist}{Equilibrium distribution.}
 	\item{...}{Not used.}
 }

 \value{
 	A process object inheriting from CodonSubst.
 }


 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{CodonSubst}}.
 }
