% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructPhylogeny.R
\name{phyloReconstruct}
\alias{phyloReconstruct}
\title{Phylogeny Reconstruction}
\usage{
phyloReconstruct(simu, which.result = NULL, ...)
}
\arguments{
\item{simu}{A object of class Phylosim}

\item{which.result}{Integer, determines which result should be used. This argument is only usefull if your 'runs' argument in \code{\link{createCompletePar}} contains more than one element. By default (NULL), the last result is used.}

\item{...}{Additional arguments to be passed to function \code{\link{hclust}}}
}
\value{
An object of class 'phylo'
}
\description{
reconstructs a phylogeny from a neutral trait matrix as returned by \code{\link{runSimulation}}
}
\examples{
# Run a simulation
par <- createCompletePar(x = 50, y = 50, dispersal = "global" , runs = 500,
        density = 1, environment = 0.5, specRate = 1)

simu <- runSimulation(par)

## Reconstruct phylogeny
rePhyl <- phyloReconstruct(simu)

## Compare to the real phylogeny (requires package 'ape')
par(mfrow=c(1,2))
plot(ape::drop.fossil(simu$Output[[1]]$phylogeny))
plot(rePhyl)
}
