% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPhylosim.R
\name{plot.PhyloSim}
\alias{plot.PhyloSim}
\title{Plot objects of class PhyloSim}
\usage{
\method{plot}{PhyloSim}(simu)
}
\arguments{
\item{simu}{Simulation output of \code{\link{runSimulation}}}
}
\description{
Plots the phylogeny tree ,the spatially abundance of the evolved species and their traits.
}
\details{
The tiplabels of the phylogeny tree are colored in correspondency to the spatial abundance plot
}
