% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPhylosimList.R
\name{plot.PhylosimList}
\alias{plot.PhylosimList}
\title{Plot all objects in an object of class PhylosimList}
\usage{
\method{plot}{PhylosimList}(simu)
}
\arguments{
\item{simu}{Simulation output of \code{\link{runSimulationBatch}}}
}
\description{
Plots the phylogeny tree ,the spatially abundance of the evolved species and their traits.
}
\details{
The tiplabels of the phylogeny tree are colored in correspondency to the spatial abundance plot
}

