% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialPhylo.R
\name{plotSpatialPhylo}
\alias{plotSpatialPhylo}
\title{Phylogeny And Spatial Abundance Plot}
\usage{
plotSpatialPhylo(simu, plot = "both", plotTraits = T, which.result = NULL)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs to contain at least the spatial species matrix ($specMat) and the phylogeny ($phylogeny)}

\item{plot}{defines what to plot."spatial" plots the spatial abundance, "phylogeny" the phylogeny tree. "both" plots both of them side by side}

\item{plotTraits}{logical, defines if the traits should be plotted}

\item{which.result}{Integer, determines which result should be used. This argument is only usefull if your 'runs' argument in \code{\link{createCompletePar}} contains more than one element. By default (NULL), the last result is used.}
}
\description{
Plots the phylogeny tree ,the spatially abundance of the evolved species and their traits.
}
\details{
The tiplabels of the phylogeny tree are colored in correspondency to the spatial abundance plot
}
\examples{

#Load data

data(simu.neutral.global)
data(simu.neutral.local)
data(simu.envcom.local)

#neutral model with global dispersal
simu <- simu.neutral.global
   #phylogeny and traits
   plotSpatialPhylo(simu, plot = "phylogeny")    

   #phylogeny and spatial distribution
   plotSpatialPhylo(simu, plot = "both", plotTraits = FALSE)
   
   #all three
   plotSpatialPhylo(simu, plot = "both", plotTraits = TRUE)

 
 #neutral model with local dispersal
 simu <- simu.neutral.local
   #phylogeny and traits
   plotSpatialPhylo(simu, plot = "phylogeny")    

   #phylogeny and spatial distribution
   plotSpatialPhylo(simu, plot = "both", plotTraits = FALSE)
   
   #all three
   plotSpatialPhylo(simu, plot = "both", plotTraits = TRUE)
   

 #evironment and competition model with local dispersal
 simu <- simu.envcom.local
   #phylogeny and traits
   plotSpatialPhylo(simu, plot = "phylogeny")    

   #phylogeny and spatial distribution
   plotSpatialPhylo(simu, plot = "both", plotTraits = FALSE)
   
   #all three
   plotSpatialPhylo(simu, plot = "both", plotTraits = TRUE)

#Set simulation run with which.result
#Here, two simulation runs are stored in simu (one with 500 timesteps, one with 1000)
par <- createCompletePar(x = 50, y = 50, dispersal = 1, runs = c(500,1000), density = 1, environment = 0.5, specRate = 1)
simu <- runSimulation(par)

#Choose to take the first simulation run (default is the last)
plotSpatialPhylo(simu, which.result = 1)


}
