% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{prepareForAnalysis}
\alias{prepareForAnalysis}
\title{Prepare for Analysis}
\usage{
prepareForAnalysis(dir = ".")
}
\arguments{
\item{dir}{root directory of the analysis (as specified in runPhylosimAnalysis)}
}
\value{
Create a subfolder "analysis_ready" and write the files "parameters" (variable parameters), "summaries" and "summaries_scaled" containing data.frames with the corresponding data.
}
\description{
Reads the files generated by combineChunks and creates dataframe for later analysis.
}
\author{
Tankred Ott
}
