% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPhylosimList.R
\name{print.PhylosimList}
\alias{print.PhylosimList}
\title{Prints summary for objects of class PhyloSim}
\usage{
\method{print}{PhylosimList}(simu)
}
\arguments{
\item{simu}{Simulation output of \code{\link{runSimulation}}}
}
\description{
Prints parameter settings, runtime and species Richness.
}

