%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{print.PSRootSummary}
\alias{print.PSRootSummary}
\alias{PSRootSummary.print}
\alias{print,PSRootSummary-method}

 \title{Print out a PSRootSummary object}

 \description{
	Print out a PSRootSummary object.
 }

 \usage{
\method{print}{PSRootSummary}(x, ...)
}

 \arguments{
 	\item{x}{A PSRootSummary object.}
 	\item{...}{Not used.}
 }

 \value{
 	The summary object (invisible).
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{PSRootSummary}}.
 }
