%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{PSRootSummary}
\docType{class}
\alias{PSRootSummary}


 \title{The PSRootSummary class}

 \description{
	PSRootSummary objects are blessed lists containing summary entries created by
	\code{summary.*} methods.

	Package:   \cr
\bold{Class PSRootSummary}\cr

\code{+--}\code{PSRootSummary}\cr

\bold{Directly known subclasses:}\cr
\cr

public class \bold{PSRootSummary}\cr


 }

 \usage{
PSRootSummary(summary=list(), ...)
}

 \arguments{
 	\item{summary}{A list.}
 	\item{...}{Not used.}
	}

 \section{Fields and Methods}{
 	\bold{Methods:}\cr
\tabular{rll}{
 \tab \code{checkConsistency} \tab  -\cr
 \tab \code{print} \tab  -\cr
}


 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{PSRootSummary}}.
 }
