% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{rejectionSamplePhylosimAnalysis}
\alias{rejectionSamplePhylosimAnalysis}
\title{Rejection sampling for Phylosim}
\usage{
rejectionSamplePhylosimAnalysis(p, s, nDraws, eQuantile = 0.025,
  nCores = "auto", which = c("none", "saABC", "rf"), sumFun = median, ...)
}
\arguments{
\item{p}{Path to a file containing the parameters in form of a data.frame/matrix, or a matrix/data.frame containing the parameters.}

\item{s}{Path to a file containing the (scaled) summary statistics, or a matrix/data.frame containing the summaries.}

\item{nDraws}{Number of ABC rejection sampling draws}

\item{eQuantile}{Epsilon expressed as quantile. E.g. 0.025 are only the nearest 2.5\%.}

\item{nCores}{Number of parallel cores, or "auto".}

\item{which}{vector of strings, determining which summary statistics reduction should be applied. Each supplied method will be applied separately. Possible options: "none", "saABC", "rf"}

\item{sumFun}{Function to summarize the accepted simulated parameters, e.g. mean or median}

\item{...}{additional parameters passed to summaryRedRF and summaryRedsaABC}
}
\value{
A list of matrices, each containing the true parameters postfixed with "_t" and the estimated parameters postfixed with "_e".
}
\description{
Applies rejection sampling using parameter and summary files generated by the PhyloSim analysis pipeline (runPhylosimAnalysis, etc.).
}
\author{
Tankred Ott
}
