% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{runPhylosimAnalysis}
\alias{runPhylosimAnalysis}
\title{Run PhyloSim Analysis}
\usage{
runPhylosimAnalysis(
  variableParams,
  fixedParams,
  nRuns,
  outDir,
  summariesOnly = FALSE,
  nCores = "auto",
  chunkSize = "auto"
)
}
\arguments{
\item{variableParams}{A named list, where the names are parameter names and the values are priors (functions). A call to the function must return a single value. See details.}

\item{fixedParams}{A named list, where the names are parameter names and the values are parameter values for the PhyloSim model.}

\item{nRuns}{Integer, specifying the number of simulations to run}

\item{outDir}{String, specifying the directory to which the output will be written}

\item{summariesOnly}{Boolean, specifying whether only the summary statistics should be saved.}

\item{nCores}{Integer or "auto", number of parallel cores to use for simulations}

\item{chunkSize}{Integer, specifying the number of simulations per save file. Choosing a very large chunkSize (>30,000) might deplete the RAM, leading to bad performance.}
}
\description{
This function is used to perform many (>1000) PhyloSim simulations in parallel with only one function call. The user can specify parameters that are fixed across all simulations and parameters that are variable. The results are saved to a specified folder.
}
\details{
variableParams must be list of functions. The idea behind this is to draw from a prior, so the functions should return a single value drawn from the corresponding parameter prior. However, the implementation allows the user to specify stateful function, instead of priors to perform analyses with variable Parameters that are not drawn from a prior, but follow a sequence e.g. run simulations with varying specRate.
}
\examples{
variableParams <- list(environment=function()return(runif(1)),
                       specRate=function()return(runif(1, 1, 10)))
fixedParams <- list(x=30, y=30,
                    runs=1000,
                    densityCut=1,
                    calculateSummaries=TRUE

runPhylosimAnalysis(variableParams,
                    fixedParams,
                    1000,
                    "outDirName")

}
\author{
Tankred Ott
}
