% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulationBatch.R
\name{runSimulationBatch}
\alias{runSimulationBatch}
\title{Batch runner}
\usage{
runSimulationBatch(pars, parallel = FALSE, backup = FALSE)
}
\arguments{
\item{pars}{A list of parameter sets as created by \code{\link{createCompletePar}}}

\item{parallel}{Integer, determining the number of cores used for parallel computing. If parallel = "auto", n-1 cores are used.}

\item{backup}{Logical, determining whether the results of the individual scenario runs should be saved as a workspace image (advised if the simulation takes a long time, or if the individual scenarios vary greatly in runtime. Default is FALSE)}
}
\value{
An object of class "PhylosimList".
}
\description{
A model of species community assembly under different assembly mechanisms, using parallel computing to make use of multi core CPUs and clusters in order to reduce computation time. The function is an extension to \link{runSimulation} in order to accelerate the simulation of multiple scenarios.
}
\details{
The "PhylosimList" object is a list of "Phylosim" objects. They can be accessed by indexing (see Example).\cr\cr This function uses the \code{\link{foreach}} and \code{\link{doParallel}} package to compute the model scenarios parallel on several cores. \cr\cr The phylogeny is passed to R in the newick format and parsed to an object of class "phylo" with the function \code{\link[ape]{read.tree}} from the \code{\link{ape}} package.
}
\examples{
## Define two parameter sets
par1 <- createCompletePar(x = 50, y = 50, dispersal = 0 , runs = c(500,1000),
        density = 0, environment = 0.5, specRate = 1) 
par2 <- createCompletePar(x = 50, y = 50, dispersal = 0 , runs = c(500,1000),
        density = 1, environment = 0.5, specRate = 1)

## Merge the parameter sets. It is important to note, that the funktion
## needs a list of parameter sets.
par <- list(par1,par2)

## Run the model
simu <- runSimulationBatch(par, parallel = 2) 


## Compare the results, here the SAC is shown exemplarily.
par(mfrow=c(1,2))
sac(simu[[1]])
sac(simu[[2]])
}

