% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSAC.R
\name{sac}
\alias{sac}
\title{Species Area Curve}
\usage{
sac(
  simu,
  which.result = NULL,
  size = NULL,
  rep = 50,
  plot = T,
  title = "SAC",
  nested = FALSE
)
}
\arguments{
\item{simu}{Simulation output of the class "Phylosim", usually consisting out of several lists. Needs at least the spatial distribution of the species stored in a matrix ($specMat)}

\item{which.result}{Integer, determines which result should be used. This argument is only usefull if your 'runs' argument in \code{\link{createCompletePar}} contains more than one element. By default (NULL), the last result is used. If you choose "all" all results are shown in one plot (see 'Details).}

\item{size}{a single value or a vector determining the edge length of the subplots. If not provided, 10 logarithmic plotsizes will be used.}

\item{rep}{The number of repetitions per size to calculate the mean}

\item{plot}{Logical determining whether to plot the SAC or not}

\item{title}{String, determining the title of the plot}

\item{nested}{Logical, determining whether nested subplots should be used.}
}
\value{
A list containing the mean species richness for each size and the respective standard deviation. If which.result = "all" only the plot will be returned.
}
\description{
Plots the species area curve for a given community.
}
\details{
Displays the accumulated species richness as a function of plot size or the amount of equally sized plots. It serves as an indicator for the clustering of a species community. A positively bent curve usually indicates clustering since an increase in plot size or number leads to an increase in species richness while a negatively bent curve indicates a more neutral distribution of species within the community. \cr\cr If which.result = "all" all intermediate results are shown in one plot. The colors of the lines are plotted as a gradient from blue (first results) to red (end result).
}
\examples{

#Load data
data(simu.neutral.global)
data(simu.neutral.local)
data(simu.envcom.local)


  
#Plot SAC curve for neutral model and global dispersion

sac(simu = simu.neutral.global)

 
 
#Plot SAC curve for neutral model and local dispersion

sac(simu = simu.neutral.local)

 
 
#Plot SAC curve for environment and competition model and local dispersion

sac(simu = simu.envcom.local)

 
 
#Plot SAC curve with random plotsize 

sac(simu=simu.envcom.local, size = sort(sample(c(10:1000), size = 10)))

 
 
#Plot SAC curve with different repititions
simu <- simu.envcom.local
par(mfrow=c(3,1))
sac(simu=simu.envcom.local, rep = 3)
sac(simu=simu.envcom.local, rep = 30)
sac(simu=simu.envcom.local, rep = 30)

}
