%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setBaseFreqs.K80}
\alias{setBaseFreqs.K80}
\alias{K80.setBaseFreqs}
\alias{setBaseFreqs,K80-method}

 \title{Forbidden action: setting the base frequency parameters for a K80 object}

 \description{
       Forbidden action: setting the base frequency parameters for a K80 object.
 }

 \usage{
\method{setBaseFreqs}{K80}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{K80}}.
 }
