%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setBaseFreqs.K81}
\alias{setBaseFreqs.K81}
\alias{K81.setBaseFreqs}
\alias{setBaseFreqs,K81-method}

 \title{Forbidden action: setting the base frequency parameters for a K81 model}

 \description{
       Forbidden action: setting the base frequency parameters for a K81 model.
 }

 \usage{
\method{setBaseFreqs}{K81}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{K81}}.
 }
