%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setBigRate.Sequence}
\alias{setBigRate.Sequence}
\alias{Sequence.setBigRate}
\alias{setBigRate,Sequence-method}

 \title{Forbidden action: setting the sum of total active event rates for a Sequence object}

 \description{
       Forbidden action: setting the sum of total active event rates for a Sequence object.
 }

 \usage{
\method{setBigRate}{Sequence}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Sequence}}.
 }
