%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setBranchEvents.PhyloSim}
\alias{setBranchEvents.PhyloSim}
\alias{PhyloSim.setBranchEvents}
\alias{setBranchEvents,PhyloSim-method}

 \title{Forbidden action: setting the list of events having per-branch statistics recorded}

 \description{
       Forbidden action: setting the list of events having per-branch statistics recorded.
 }

 \usage{
\method{setBranchEvents}{PhyloSim}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{PhyloSim}}.
 }
