%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setCumulativeRates.Sequence}
\alias{setCumulativeRates.Sequence}
\alias{Sequence.setCumulativeRates}
\alias{setCumulativeRates,Sequence-method}

 \title{Forbidden action: setting the cumulative rates for the sites aggregated by a Sequence object}

 \description{
       Forbidden action: setting the cumulative rates for the sites aggregated by a Sequence object.
 }

 \usage{
\method{setCumulativeRates}{Sequence}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Sequence}}.
 }
