%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setEvents.Site}
\alias{setEvents.Site}
\alias{Site.setEvents}
\alias{setEvents,Site-method}

 \title{Forbidden action: setting the list of active events for a Site object}

 \description{
       Forbidden action: setting the list of active events for a Site object.
 }

 \usage{
\method{setEvents}{Site}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Site}}.
 }
