%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setId.PhyloSim}
\alias{setId.PhyloSim}
\alias{PhyloSim.setId}
\alias{setId,PhyloSim-method}

 \title{Forbidden action: setting the unique identifier of a PhyloSim object}

 \description{
       Forbidden action: setting the unique identifier of a PhyloSim object.

 }

 \usage{
\method{setId}{PhyloSim}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{PhyloSim}}.
 }
