%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setId.Process}
\alias{setId.Process}
\alias{Process.setId}
\alias{setId,Process-method}

 \title{Forbidden action: setting the unique Process object identifier}

 \description{
       Forbidden action: setting the unique Process object identifier.
 }

 \usage{
\method{setId}{Process}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Process}}.
 }
