%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setLength.Sequence}
\alias{setLength.Sequence}
\alias{Sequence.setLength}
\alias{setLength,Sequence-method}

 \title{Forbidden action: setting the length of a Sequence object}

 \description{
       Forbidden action: setting the length of a Sequence object.
	The length of the Sequence objects can be specified when the object is constructed,
	or modified later by the "insertSequence" and "deleteSubSequence" methods.
 }

 \usage{
\method{setLength}{Sequence}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Sequence}}.
 }
