%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setMatrix.QMatrix}
\alias{setMatrix.QMatrix}
\alias{QMatrix.setMatrix}
\alias{setMatrix,QMatrix-method}

 \title{Forbidden action: setting the unscaled rate matrix for a QMatrix object}

 \description{
       Forbidden action: setting the unscaled rate matrix for a QMatrix object.
 }

 \usage{
\method{setMatrix}{QMatrix}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{QMatrix}}.
 }
