%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setMethodsList.PSRoot}
\alias{setMethodsList.PSRoot}
\alias{PSRoot.setMethodsList}
\alias{setMethodsList,PSRoot-method}

 \title{Forbidden action: setting the list of applicable methods for an object}

 \description{
       Forbidden action: setting the list of applicable methods for an object.
 }

 \usage{
\method{setMethodsList}{PSRoot}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{PSRoot}}.
 }
