%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setSiteSpecificParamIds.Process}
\alias{setSiteSpecificParamIds.Process}
\alias{Process.setSiteSpecificParamIds}
\alias{setSiteSpecificParamIds,Process-method}

 \title{Forbidden action: setting the paramter identifiers of the site specific paramters from a Process object}

 \description{
       Forbidden action: setting the paramter identifiers of the site specific paramters from a Process object.
 }

 \usage{
\method{setSiteSpecificParamIds}{Process}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Process}}.
 }
