%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setSiteSpecificParamList.Process}
\alias{setSiteSpecificParamList.Process}
\alias{Process.setSiteSpecificParamList}
\alias{setSiteSpecificParamList,Process-method}

 \title{Forbidden action: setting the site specific paramter list for a Process object}

 \description{
       Forbidden action: setting the site specific paramter list for a Process object.
	Use .addSiteSpecificParameter to add new site specific paramters when implementing new processes.
 }

 \usage{
\method{setSiteSpecificParamList}{Process}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Process}}.
 }
