%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setSymbolLength.Alphabet}
\alias{setSymbolLength.Alphabet}
\alias{Alphabet.setSymbolLength}
\alias{setSymbolLength,Alphabet-method}

 \title{Forbidden action: setting the symbol length for an Alphabet object}

 \description{
	Forbidden action: setting the symbol length for an Alphabet object.
 }

 \usage{
\method{setSymbolLength}{Alphabet}(this, value, ...)
}

 \arguments{
 	\item{this}{An Alphabet object.}
 	\item{value}{Not used.}
 	\item{...}{Not used.}
 }

 \value{
 	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{Alphabet}}.
 }
