%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setTipLabels.PhyloSim}
\alias{setTipLabels.PhyloSim}
\alias{PhyloSim.setTipLabels}
\alias{setTipLabels,PhyloSim-method}

 \title{Forbidden action: setting the tip labels for a phylo object aggregated by a PhyloSim object}

 \description{
       Forbidden action: setting the tip labels for a phylo object aggregated by a PhyloSim object.
 }

 \usage{
\method{setTipLabels}{PhyloSim}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{PhyloSim}}.
 }
