%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setTips.PhyloSim}
\alias{setTips.PhyloSim}
\alias{PhyloSim.setTips}
\alias{setTips,PhyloSim-method}

 \title{Forbidden action: setting the node identifiers of the tip nodes for a PhyloSim object}

 \description{
       Forbidden action: setting the node identifiers of the tip nodes for a PhyloSim object.
 }

 \usage{
\method{setTips}{PhyloSim}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{PhyloSim}}.
 }
