%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setTransTable.CodonAlphabet}
\alias{setTransTable.CodonAlphabet}
\alias{CodonAlphabet.setTransTable}
\alias{setTransTable,CodonAlphabet-method}

 \title{Forbidden action: setting the genetic code table for a CodonAlphabet object}

 \description{
       Forbidden action: setting the genetic code table for a CodonAlphabet object.
	Use the \code{table.id} constructor argument to set the genetic code.
 }

 \usage{
\method{setTransTable}{CodonAlphabet}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{CodonAlphabet}}.
 }
