%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setType.BrownianInsertor}
\alias{setType.BrownianInsertor}
\alias{BrownianInsertor.setType}
\alias{setType,BrownianInsertor-method}

 \title{Forbidden action: setting the type of a BrownianInsertor object}

 \description{
       Forbidden action: setting the type of a BrownianInsertor object.

	The type can be set only from the \code{type} constructor argument and cannot be changed later.
 }

 \usage{
\method{setType}{BrownianInsertor}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{BrownianInsertor}}.
 }
