%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setType.FastFieldDeletor}
\alias{setType.FastFieldDeletor}
\alias{FastFieldDeletor.setType}
\alias{setType,FastFieldDeletor-method}

 \title{Forbidden action: setting the type of a FastFieldDeletor object}

 \description{
       Forbidden action: setting the type of a FastFieldDeletor object.

	The type can be set only through the \code{type} constructor argument.
 }

 \usage{
\method{setType}{FastFieldDeletor}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{FastFieldDeletor}}.
 }
