%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setUniqueAlphabets.Sequence}
\alias{setUniqueAlphabets.Sequence}
\alias{Sequence.setUniqueAlphabets}
\alias{setUniqueAlphabets,Sequence-method}

 \title{Forbidden action: setting the list of unique Alphabet objects attached to the Site object aggregated by a Sequence object}

 \description{
       Forbidden action: setting the list of unique Alphabet objects attached to the Site object aggregated by a Sequence object.
 }

 \usage{
\method{setUniqueAlphabets}{Sequence}(this, value, ...)
}

 \arguments{
       \item{this}{An object.}
       \item{value}{Not used.}
       \item{...}{Not used.}
 }

 \value{
	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
       For more information see \code{\link{Sequence}}.
 }
