% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data documentation.R
\name{simu.neutral.local}
\alias{simu.neutral.local}
\title{Simulation output of community assembly with local dispersion under neutral conditions}
\format{
A list containing two lists: The model output and the model input. The model output is described here
\describe{
\item{specMat}{species matrix, a matrix representing the species abundance}
\item{traitMat}{trait matrix, a matrix containing the environmental trait for each species (corresponding to the species matrix)}
\item{envMat}{environment matrix, contains the environmental conditions for each cell}
\item{compMat}{competition matrix, contains the competition attribute for each cell}
\item{neutMat}{neutral matrix, contains the neutral traits}
\item{phylogeny}{phylogeny of the community stored in a list}
}
}
\source{
created with \code{\link{runSimulation}}
}
\description{
Simulation output of community assembly with local dispersion under neutral conditions
}
