% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRichness.R
\name{specRich}
\alias{specRich}
\title{Species richness}
\usage{
specRich(simu, which.result = NULL)
}
\arguments{
\item{simu}{Object of class "Phylosim", usually consisting out of several lists. Needs at least the spatial distribution of the species stored in a matrix ($specMat)}

\item{which.result}{Integer, determines which result should be used. This argument is only usefull if your 'runs' argument in \code{\link{createCompletePar}} contains more than one element. By default (NULL), the last result is used.}
}
\value{
An integer value for the species richness
}
\description{
Calculate the total species richness of a community
}
