% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPhylosim.R
\name{summary.PhyloSim}
\alias{summary.PhyloSim}
\title{Summary for objects of class PhyloSim}
\usage{
\method{summary}{PhyloSim}(simu)
}
\arguments{
\item{simu}{Simulation output of \code{\link{runSimulation}}}
}
\description{
Gives summary about parameter settings, runtime and species Richness.
}

