% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{summaryRedRF}
\alias{summaryRedRF}
\title{random forest parameter prediction}
\usage{
summaryRedRF(p, s, nCores = "auto", nTrees = 500, scale = TRUE, ...)
}
\arguments{
\item{p}{matrix or dataframe containing the parameters as columns}

\item{s}{matrix or dataframe containing the summary statistics as columns}

\item{nCores}{number of parallel cores}

\item{nTrees}{number of trees to construct}

\item{scale}{logical, determining whether the resulting predictions should be scaled by their variance}

\item{...}{further arguments passed to ranger::ranger}
}
\value{
A list with two items. Item one contains the true parameters, item two contains the predicted parameters.
}
\description{
Function to predict model parameters on the basis of summary statistics. This function is used to create a reduced set of summary statistics by fitting a random forest with p as response and s as predictors. The predicted parameters can then be used in the ABC rejection sampling function instead of the summaries.
}
\author{
Tankred Ott
}
