% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{summaryRedsaABC}
\alias{summaryRedsaABC}
\title{saABC parameter prediction}
\usage{
summaryRedsaABC(p, s, scale = TRUE, ...)
}
\arguments{
\item{p}{matrix or dataframe containing the parameters as columns}

\item{s}{matrix or dataframe containing the summary statistics as columns}

\item{scale}{logical, determining whether the resulting predictions should be scaled by their variance}

\item{...}{further arguments passed to abctools::saABC}
}
\value{
Matrix of parameter predictions
}
\description{
Function to predict model parameters on the basis of summary statistics. This function is used to create a reduced set of summary statistics using saABC with p as response and s as predictors. The predicted parameters can then be used in the ABC rejection sampling function instead of the summaries.
}
\author{
Tankred Ott
}
