% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus_in.R
\name{torus_in}
\alias{torus_in}
\title{Internal torus extension}
\usage{
torus_in(matrix, max_neighborhood_radius)
}
\arguments{
\item{matrix}{A 2D numeric or integer matrix}

\item{max_neighborhood_radius}{Radius of the toroidal wrap}
}
\value{
A matrix enlarged by 2*r in both dimensions
}
\description{
Adds toroidal border of radius \code{r} to a 2D matrix.
The resulting matrix has \code{dim + 2*r} rows and columns.
}
\details{
Expand matrix with toroidal border
}
\keyword{internal}
