% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{treeMetrics}
\alias{treeMetrics}
\title{Calculate summary statistics from a phylogenetic tree}
\usage{
treeMetrics(phy)
}
\arguments{
\item{object}{an object of class phylo}
}
\description{
Calculate summary statistics from a phylogenetic tree
}
\details{
See also 

https://www.lirmm.fr/MEP05/talk/14_Mooers.pdf

https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5790134/
}
\author{
Florian Hartig, modified from a function by Allen Hulbert
}
