%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Undocumented.PhyloSim}
\alias{Undocumented.PhyloSim}
\alias{PhyloSim.Undocumented}
\alias{newMatrix}
\alias{setEquDist.CodonSubst}
\alias{Undocumented,PhyloSim-method}
 \alias{Undocumented}
 \alias{BrownianPath}
 \alias{buildFromPAML}
 \alias{checkConsistency}
 \alias{clearStates}
 \alias{areSynonymous}
 \alias{attachHookToNode}
 \alias{attachProcess}
 \alias{attachSeqToNode}
 \alias{copySubSequence}
 \alias{Debug}
 \alias{deleteSubSequence}
 \alias{detachHookFromNode}
 \alias{detachProcess}
 \alias{detachSeqFromNode}
 \alias{enableVirtual}
 \alias{exportStatTree}
 \alias{flagTotalRate}
 \alias{generateInsert}
 \alias{getAcceptBy}
 \alias{getAcceptWin}
 \alias{getAlignment}
 \alias{getAlphabet}
 \alias{getAlphabets}
 \alias{getAlphabet}
 \alias{getAncestral}
 \alias{getAncestral}
 \alias{getBaseFreqs}
 \alias{getBigRate}
 \alias{getBranchEvents}
 \alias{getCodonFreqs}
 \alias{getComments}
 \alias{getCumulativeRates}
 \alias{getCumulativeRatesFromRange}
 \alias{getDeletionTolerance}
 \alias{getDist}
 \alias{getEdge}
 \alias{getEdges}
 \alias{getEquDist}
 \alias{getEventRate}
 \alias{getEventRateAtSite}
 \alias{getEvents}
 \alias{getEventsAtSite}
 \alias{getGenerateBy}
 \alias{getHandler}
 \alias{getId}
 \alias{getInsertHook}
 \alias{getInsertionTolerance}
 \alias{getInsertionTolerance}
 \alias{getKappa}
 \alias{getLength}
 \alias{getLengthParam1}
 \alias{getLengthParam2}
 \alias{getLogFile}
 \alias{getLogLevel}
 \alias{getMatrix}
 \alias{getMaxLength}
 \alias{getMethodsList}
 \alias{getNedges}
 \alias{getNodes}
 \alias{getNtips}
 \alias{getOmegas}
 \alias{getParameterAtSite}
 \alias{getParameterAtSites}
 \alias{getPhylo}
 \alias{getProbs}
 \alias{getProcess}
 \alias{getProcesses}
 \alias{getProposeBy}
 \alias{getQMatrix}
 \alias{getRate}
 \alias{getRateList}
 \alias{getRateMultipliers}
 \alias{getRateParam}
 \alias{getRateParamList}
 \alias{getRootNode}
 \alias{getRootSeq}
 \alias{getScale}
 \alias{getScaledMatrix}
 \alias{getSeqFromNode}
 \alias{getSequence}
 \alias{getSequences}
 \alias{getSite}
 \alias{getSites}
 \alias{getSiteSpecificParamIds}
 \alias{getSiteSpecificParamList}
 \alias{getSize}
 \alias{getSizes}
 \alias{getState}
 \alias{getStates}
 \alias{getString}
 \alias{getSymbolFreqs}
 \alias{getSymbolLength}
 \alias{getSymbols}
 \alias{getTableId}
 \alias{getTargetState}
 \alias{getTemplateSeq}
 \alias{getTheta}
 \alias{getTipLabels}
 \alias{getTips}
 \alias{getToleranceMargin}
 \alias{getTotalRate}
 \alias{getTotalRates}
 \alias{getTotalRatesFromRange}
 \alias{getTransTable}
 \alias{getTreeLength}
 \alias{getType}
 \alias{getUniqueAlphabets}
 \alias{getUniqueProcesses}
 \alias{getWriteProtected}
 \alias{globalConsistencyCheck}
 \alias{hasSiteSpecificParameter}
 \alias{hasSymbols}
 \alias{hasUndefinedRate}
 \alias{insertSequence}
 \alias{intersect}
 \alias{list}
 \alias{isAttached}
 \alias{isEmpty}
 \alias{is}
 \alias{default}
 \alias{isStartCodon}
 \alias{isStopCodon}
 \alias{is}
 \alias{tip}
 \alias{Log}
 \alias{my}
 \alias{all}
 \alias{equal}
 \alias{newAAMatrix}
 \alias{omegaHist}
 \alias{omegaVarM0}
 \alias{omegaVarM1}
 \alias{omegaVarM10Cont}
 \alias{omegaVarM10Cont}
 \alias{omegaVarM2}
 \alias{omegaVarM3}
 \alias{omegaVarM4}
 \alias{Perform}
 \alias{plotParametersAtSites}
 \alias{plot}
 \alias{plusGamma}
 \alias{plusInvGamma}
 \alias{proposeLength}
 \alias{rescaleQMatrix}
 \alias{sampleState}
 \alias{sampleStates}
 \alias{saveAlignment}
 \alias{Scale}
 \alias{scaleTree}
 \alias{setAcceptBy}
 \alias{setAcceptWin}
 \alias{setAlignment}
 \alias{setAlphabet}
 \alias{setAlphabets}
 \alias{setAncestral}
 \alias{setBaseFreqs}
 \alias{setBigRate}
 \alias{setBranchEvents}
 \alias{setCodonFreqs}
 \alias{setCodonFreqs}
 \alias{setComments}
 \alias{setCumulativeRates}
 \alias{setDeletionTolerance}
 \alias{setDist}
 \alias{setEdges}
 \alias{setEquDist}
 \alias{setEquDist}
 \alias{setEvents}
 \alias{setGenerateBy}
 \alias{setHandler}
 \alias{setId}
 \alias{setInsertHook}
 \alias{setInsertionTolerance}
 \alias{setInsertionTolerance}
 \alias{setKappa}
 \alias{setKappa}
 \alias{setLength}
 \alias{setLengthParam1}
 \alias{setLengthParam2}
 \alias{setLogFile}
 \alias{setLogLevel}
 \alias{setMatrix}
 \alias{setMaxLength}
 \alias{setMethodsList}
 \alias{setName}
 \alias{setNedges}
 \alias{setNodes}
 \alias{setNtips}
 \alias{setOmegas}
 \alias{setParameterAtSite}
 \alias{setParameterAtSites}
 \alias{setPhylo}
 \alias{setPosition}
 \alias{setProbs}
 \alias{setProcess}
 \alias{setProcesses}
 \alias{setProposeBy}
 \alias{setQMatrix}
 \alias{setRate}
 \alias{setRateList}
 \alias{setRateMultipliers}
 \alias{setRateParam}
 \alias{setRateParamList}
 \alias{setRootNode}
 \alias{setRootSeq}
 \alias{setScale}
 \alias{setScaledMatrix}
 \alias{setSequence}
 \alias{setSequences}
 \alias{setSite}
 \alias{setSiteSpecificParamIds}
 \alias{setSiteSpecificParamList}
 \alias{setSize}
 \alias{setSizes}
 \alias{setState}
 \alias{setStates}
 \alias{setString}
 \alias{setSymbolLength}
 \alias{setSymbols}
 \alias{setTableId}
 \alias{setTargetState}
 \alias{setTemplateSeq}
 \alias{setTheta}
 \alias{setTipLabels}
 \alias{setTips}
 \alias{setToleranceMargin}
 \alias{setTotalRate}
 \alias{setTotalRates}
 \alias{setTransTable}
 \alias{setTreeLength}
 \alias{setType}
 \alias{setUniqueAlphabets}
 \alias{setUniqueProcesses}
 \alias{setWriteProtected}
 \alias{Simulate}
 \alias{Translate}
 \alias{translateCodon}
 \alias{virtualAssignmentForbidden}
 \alias{intersect.list}
 \alias{is.tip}
 \alias{my.all.equal}
 \alias{plot.PSRoot}
 \alias{revComp}
 \alias{readAlignment}
 \alias{readTree}
 \alias{getOmegaScalingFactor}
 \alias{saveLoadReference}
 \alias{getAlignmentLength}

 \title{Undocumented object (PhyloSim package)}

 \description{
	Undocumented object (PhyloSim package).

	See the corresponding specific methods if applicable.
 }

 \usage{
\method{Undocumented}{PhyloSim}(...)
}

 \arguments{
 	\item{...}{Not used.}
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{PhyloSim}}.
 }
