%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ../PhyloSimSource.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{virtualAssignmentForbidden.PSRoot}
\alias{virtualAssignmentForbidden.PSRoot}
\alias{PSRoot.virtualAssignmentForbidden}
\alias{virtualAssignmentForbidden,PSRoot-method}

 \title{Throws an error message informing the user about forbidden action on virtual a field}

 \description{
	Throws an error message informing the user about forbidden action on virtual a field.
 }

 \usage{
\method{virtualAssignmentForbidden}{PSRoot}(this, ...)
}

 \arguments{
 	\item{this}{A PSRoot object.}
 	\item{...}{Not used.}
 }

 \value{
 	Throws an error.
 }

 \author{Botond Sipos, Gregory Jordan}

 \seealso{
 	For more information see \code{\link{PSRoot}}.
 }
