% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{calculateSummaryStatistics}
\alias{calculateSummaryStatistics}
\title{Summary Statistics}
\usage{
calculateSummaryStatistics(simulation)
}
\arguments{
\item{simulation}{Simulation output of the class "Phylosim"}
}
\value{
A list containing the summary statistics.
}
\description{
Calculates the summary satistics for a simulation.
}
\details{
TODO: implement
}
\examples{

simu <- runSimulation(createCompletePar(x=50, y=50, runs=1000, specRate=2))
summStat <- calculateSummaryStatistics(simu)


}
