% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{callModel}
\alias{callModel}
\title{Core phylosim model}
\usage{
callModel(
  x,
  y,
  dispersal,
  runs,
  specRate,
  dens,
  env,
  neutral,
  mort,
  mortStrength,
  repro,
  dispersalCutoff,
  densityCutoff,
  seed,
  envStrength,
  compStrength,
  fission,
  redQueen,
  redQueenStrength,
  protracted,
  airmatR,
  soilmatR,
  prunePhylogeny
)
}
\description{
Core phylosim model
}
