% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collessImbalance.R
\name{collessImbalance}
\alias{collessImbalance}
\title{Colless' imbalance}
\usage{
collessImbalance(simu, which.simulation = NULL)
}
\arguments{
\item{simu}{An object of type "PhyloSim"}

\item{which.simulation}{Integer, determines which simulation should be used. Only useful if your Phylosim object contains more than one result. By default the last result is chosen.}
}
\value{
A numeric value for the Colless' Imbalance
}
\description{
Calculates the Colless' imbalance for a Phylogeny.
}

