% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collessImbalance.R
\name{collessImbalance}
\alias{collessImbalance}
\title{Colless' imbalance}
\usage{
collessImbalance(simu, which.result = NULL)
}
\arguments{
\item{simu}{An object of type "PhyloSim"}

\item{which.result}{Integer, determines which result should be used. This argument is only usefull if your 'runs' argument in \code{\link{createCompletePar}} contains more than one element. By default (NULL), the last result is used.}
}
\value{
A numeric value for the Colless' Imbalance
}
\description{
Calculates the Colless' imbalance for a Phylogeny.
}

