% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosimAnalysis.R
\name{rejectionSamplePhylosimAnalysis}
\alias{rejectionSamplePhylosimAnalysis}
\title{Rejection sampling for Phylosim}
\usage{
rejectionSamplePhylosimAnalysis(parameterFile, summaryFile, nDraws = "auto",
  eQuantile = 0.025, nCores = "auto", saABC = FALSE, sumFun = median)
}
\arguments{
\item{parameterFile}{Path to a file containing the parameters in form of a data.frame}

\item{summaryFile}{Path to a file containing the (scaled) summary statistics}

\item{nDraws}{Number of ABC rejection sampling draws}

\item{eQuantile}{Epsilon expressed as quantile. E.g. 0.025 are only the nearest 2.5\%.}

\item{nCores}{Number of parallel cores, or "auto".}

\item{saABC}{Logical, determining whether saABC should be used to reduce the number of summaries.}

\item{sumFun}{Function to summarize the accepted simulated parameters, e.g. mean or median}
}
\value{
A data.frame containing the true parameters postfixed with "_t" and the estimated parameters postfixed with "_e".
}
\description{
Applies rejection sampling using parameter and summary files generated by the PhyloSim analysis pipeline (runPhylosimAnalysis, etc.).
}
\author{
Tankred Ott
}
