% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{treeMetrics}
\alias{treeMetrics}
\title{Calculate summary statistics from a phylogenetic tree}
\usage{
treeMetrics(phy, selection = 1:10, dropFossils = FALSE)
}
\arguments{
\item{selection}{selection of metrics. 1:10 is fast, > 11 is slow. Max range is 1:13}

\item{dropFossils}{whether to apply ape::drop.fossil. Safer, but can be slow for large trees.}

\item{object}{an object of class phylo}
}
\description{
Calculate summary statistics from a phylogenetic tree
}
\details{
See also

https://www.lirmm.fr/MEP05/talk/14_Mooers.pdf

https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5790134/
}
\author{
Florian Hartig, modified from a function by Allen Hulbert
}
